requirements.txt
flask>=2.0.0
numpy>=1.21.0
joblib>=1.0.0
bleach>=5.0.0
python-dotenv>=0.19.0

# Optional - install if you have Visual C++ Build Tools
# scikit-learn>=1.0.0
